% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survreg.R
\name{broom_tidy_survreg}
\alias{broom_tidy_survreg}
\title{Broom::tidy.survreg}
\usage{
broom_tidy_survreg(x, conf.level = 0.95, ...)
}
\arguments{
\item{x}{An object to be converted into a tidy data.frame}

\item{conf.level}{confidence level for CI}

\item{...}{extra arguments (not used)}
}
\description{
Bugfixed version of \link[broom]{tidy.survreg}
}
