% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_matrix.R
\name{df.as.matrix.dummy.withfactor}
\alias{df.as.matrix.dummy.withfactor}
\title{data.frame to matrix}
\usage{
df.as.matrix.dummy.withfactor(df.x, desc.basic)
}
\arguments{
\item{df.x}{(\code{data.frame}) to be converted}

\item{desc.basic}{the descriptor returned by \code{\link{df.describe.basic}}}
}
\description{
data.frame to matrix
}
