% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_describe.R
\name{df.describe.factors}
\alias{df.describe.factors}
\title{Data frame factors description}
\usage{
df.describe.factors(df)
}
\arguments{
\item{df}{(\code{data.frame}) to be described}
}
\description{
Data frame factors description
}

