% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.R
\name{glmnet.df.as.matrix}
\alias{glmnet.df.as.matrix}
\title{Data frame to matrix conversion}
\usage{
glmnet.df.as.matrix(df.x, use.and.remove.intercept = FALSE)
}
\arguments{
\item{df.x}{(\code{data.frame}) to be encoded}

\item{use.and.remove.intercept}{if \code{TRUE} the}
}
\description{
Simple conversion of \code{data.frame}, with default factors encoding,
using the \code{\link[stats]{model.matrix}} with formula \code{(~.)} or \code{(~. - 1)},
and removal of resulting \code{intercept}
column (if any).
}
\details{
If the formula \code{(~.)} is used (\code{use.and.remove.intercept = TRUE}):
\itemize{
 \item{each factor is contrasted \emph{vs} its first level, that is}
 \item{the first level of each factor is dropped and all the remaining levels are encoded
 as the binary columns in the resulting matrix}
}

Function provides the description of the transformation, mapping the columns and factor levels to
the resulting matrix.
}

