% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.R
\name{inference.lasso.path.structural.changes}
\alias{inference.lasso.path.structural.changes}
\title{Extraction of critical points from Lasso path}
\usage{
inference.lasso.path.structural.changes(model.full.lasso, strict = FALSE)
}
\arguments{
\item{model.full.lasso}{the model fitted by glmnet::glmnet function}

\item{strict}{if \code{TRUE} only the points of the structural change will be returned}
}
\description{
Function extracts the points of the structure change (adding/dropping coefficient)
as well as the preceeding ones, along with their deviance
}

