% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_language.R, R/mixl_obj_parsing.R
\name{mixl_parse_specification}
\alias{mixl_parse_specification}
\alias{mixl_parse_specification}
\title{Model description parser}
\usage{
mixl_parse_specification(model.description.named)

mixl_parse_specification(model.description.named)
}
\arguments{
\item{model.description}{a model-level list with single model}

\item{model.description}{a model-level list with single model}
}
\description{
Model description parser

Model description parser
}
