% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_language.R
\name{mixl_specification_read_and_parse}
\alias{mixl_specification_read_and_parse}
\title{MixLang Model Parser}
\usage{
mixl_specification_read_and_parse(fpath.vec, verbose = FALSE)
}
\arguments{
\item{fpath.vec}{(\code{character()}) vector of file names}

\item{verbose}{should the function be verbose}
}
\description{
Function reads and parses model specification files
}
