% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_language_preparser.R
\name{mixl_specification_resolve_inner_preparse}
\alias{mixl_specification_resolve_inner_preparse}
\title{Extends-aware Model Parsing Order}
\usage{
mixl_specification_resolve_inner_preparse(raw_spec, verbose = FALSE)
}
\description{
Extends-aware Model Parsing Order
}
