% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.R
\name{structural.changes.col.matrix}
\alias{structural.changes.col.matrix}
\title{Extraction of critical steps from Lasso path}
\usage{
structural.changes.col.matrix(mtx, strict = FALSE)
}
\arguments{
\item{mtx}{the coefficients}

\item{strict}{if \code{TRUE} only the points of the structural change will be returned}
}
\value{
\code{(numeric)} vector containing the steps with structural changes
}
\description{
Function extracts the steps of the LASSO path resulting in the structural changes of
the model coefficients matrix (adding/dropping coefficient)
as well as the preceeding ones
}

