% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.R
\name{glmnet.process.output}
\alias{glmnet.process.output}
\title{The glmnet postprocessor}
\usage{
glmnet.process.output(glmnet.res)
}
\arguments{
\item{glmnet.res}{the result of the \code{\link[glmnet]{glmnet}} function}
}
\value{
the \code{list} with following fields:
 \itemize{
   \item{\code{steps}: a \code{data.frame} with columns:
     \itemize{
       \item{\code{step}}
       \item{\code{lambda}}
       \item{\code{dev.ratio}}
       \item{\code{df}}
       \item{\code{deviance}}
     }
   }
   \item{\code{coeffs}: a \code{data.frame} with columns:
     \itemize{
       \item{\code{step}}
       \item{\code{term}}
       \item{\code{estimate}}
     }
   }
 }
}
\description{
The function is based on the results returned by \code{\link[broom]{glmnet_tidiers}}
}

